fix(platform): 修复 timerfd sentinel 泄漏导致 SegFault + 编译警告 + 测试修复
- platform/linux/port.c: timerfd sentinel 始终过滤（去掉 timer_armed 条件），
  非 armed 路径 drain timerfd，修复 broker INFINITE wait 下 sentinel 泄漏为
  合法指针导致的 SegFault（影响 timer + 全部集成测试）
- test/test_helpers.h: Linux 分支补充 #include <errno.h>，
  test_signal_raw_waitable 加 __attribute__((unused))
- test/integration/*.c: int→unsigned int、%zu→%u、sign-conversion cast，
  消除全部 24 个编译警告
- test/unit/test_broker.c: test_watcher_deinit_before_remove 修正为
  remove→deinit 顺序，消除 ASAN 泄漏
- test/unit/test_platform_backend.c: level_triggered 测试期望修正为
  cnt>=1（waitset 不负责 drain eventfd）
- docs/audit/*: 3 份审计文档追加 P11.2 补丁记录
- docs/plans/共识计划.md: P9 处置记录追加 P11.2 条目

验证：linux-release 零警告零错误，21/21 CTest 全绿（含 20 分钟压力测试）

+++ test/integration/scenario_cocktail.c
@@ -29,7 +29,7 @@ typedef struct {
     uint8_t data[CKTL_LARGE_SIZE - 8u];
 } cktl_large_t;
 
-#define CKTL_MAGIC 0xDEADBEEF
+#define CKTL_MAGIC ((int32_t)0xDEADBEEF)
 
 static const size_t g_cktl_payload_sizes[CKTL_NUM_TYPES] = 
     {0u, sizeof(int32_t), sizeof(cktl_small_t), sizeof(cktl_large_t)};
